add Delete All

+++ src/components/Todo.js
@@ -7,34 +7,40 @@ class Todo extends React.Component {
     toDoList: [],
   };
   handleInputKeyPress = (event) => {
+    if (event.target.value === '') {
+      alert('Todo를 입력해주세요.');
+      return;
+    }
     const {
       target: { value },
     } = event;
     if (event.key === 'Enter') {
       this.setState(
         (state) => ({ toDoList: [...state.toDoList, value] }),
-        () => {
-          localStorage.setItem(TODO, JSON.stringify(this.state.toDoList));
-        }
+        () => localStorage.setItem(TODO, JSON.stringify(this.state.toDoList))
       );
       event.target.value = '';
     }
   };
 
   handleClickRemove = (index) => {
-    const { toDoList } = this.state;
-
-    this.setState(
-      (state) => ({
-        toDoList: [
-          ...state.toDoList.slice(0, index),
-          ...state.toDoList.slice(index + 1),
-        ],
-      }),
-      () => localStorage.setItem(TODO, JSON.stringify(this.state.toDoList))
-    );
+    if (window.confirm('목록에서 삭제하시겠습니까?')) {
+      this.setState(
+        (state) => ({
+          toDoList: [
+            ...state.toDoList.slice(0, index),
+            ...state.toDoList.slice(index + 1),
+          ],
+        }),
+        () => localStorage.setItem(TODO, JSON.stringify(this.state.toDoList))
+      );
+    }
   };
-
+  handleClickRemoveAll = () => {
+    if (window.confirm('목록 전체를 삭제하시겠습니까?')) {
+      this.setState({ toDoList: [] }, () => localStorage.clear());
+    }
+  };
   componentDidMount() {
     this.setState({
       toDoList: JSON.parse(localStorage.getItem(TODO)) || [],
@@ -44,6 +50,7 @@ class Todo extends React.Component {
     const { toDoList } = this.state;
     return (
       <Container>
+        <Button onClick={this.handleClickRemoveAll}>목록전체삭제</Button>
         <Input
           placeholder="할 일 추가"
           onKeyPress={this.handleInputKeyPress}
@@ -59,6 +66,17 @@ class Todo extends React.Component {
   }
 }
 
+const Button = styled.button`
+  height: 100%;
+  width: auto;
+  margin: 22px;
+  outline: none;
+  border: 1px solid silver;
+  border-radius: 11px;
+  background: transparent;
+  font-size: 15px;
+  color: white;
+`;
 const TodoContainer = styled.div`
   text-align: center;
 `;
@@ -71,6 +89,7 @@ const Container = styled.div`
 const Input = styled.input`
   width: 80%;
   height: 33px;
+
   padding: 7px;
   outline: none;
   border: 1px solid silver;
